Refactor code structure for improved readability and maintainability

+++ backend/controllers/pizza.controller.js
@@ -19,7 +19,6 @@ export const addPizza = async (req, res) => {
     }
 };
 
-
 // Function to get all pizzas
 export const getAllPizzas = async (req, res) => {
     try {
@@ -33,15 +32,51 @@ export const getAllPizzas = async (req, res) => {
 // Function to delete a pizza by ID
 export const deletePizza = async (req, res) => {
     try {
-        const pizzaId = req.body.id; // Read the id from the request body
+        const { id } = req.params;
+        const pizza = await PizzaModel.findById(id);
+
+        if (!pizza) {
+            return res.status(404).json({ message: "Pizza not found" });
+        }
+
+        // Check if the file exists before deleting
+        const filePath = `uploads/${pizza.image}`;
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        }
+
+        // Delete the pizza from the database
+        await PizzaModel.findByIdAndDelete(id);
+
+        res.status(200).json({ message: "Pizza deleted successfully" });
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
+};
+// Function to update a pizza by ID
+export const updatePizza = async (req, res) => {
+    try {
+        const pizzaId = req.params.id; 
+        const { name, description, price } = req.body;
+        const pizzaImage = req.file ? req.file.filename : null; 
+
         const pizza = await PizzaModel.findById(pizzaId);
         if (!pizza) {
             return res.status(404).json({ message: 'Pizza not found' });
         }
-        // Delete the image file from the server
-        fs.unlinkSync(`uploads/${pizza.image}`);
-        await PizzaModel.findByIdAndDelete(pizzaId);
-        res.status(200).json({ message: 'Pizza deleted successfully' });
+
+        // Update the pizza details
+        pizza.name = name || pizza.name;
+        pizza.description = description || pizza.description;
+        pizza.price = price || pizza.price;
+        if (pizzaImage) {
+            // Delete the old image file if a new one is provided
+            fs.unlinkSync(`uploads/${pizza.image}`);
+            pizza.image = pizzaImage; // Save the new image filename
+        }
+
+        await pizza.save();
+        res.status(200).json({ message: 'Pizza updated successfully' });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
